format: condensed code into one file
