--- app_test_idea/test_app_output.py
@@ -1,10 +1,0 @@
-
-import config_app
-import os
-
-app = config_app.button_prompts()
-
-
-def test_app_file_storage() -> None:
-    app.store_in_file()
-    assert os.path.isfile("checks.txt")
