+++ chasten/configApp.py
@@ -1,10 +1,45 @@
 # Import necessary modules and components from the Textual library,
 # as well as other Python modules like os and validation tools.
 import os
+from typing import List
+
 from textual.app import App, ComposeResult
-from textual.widgets import Input, Button, Pretty, Static
 from textual.validation import Number
-from config_app_ouput import ProperSentence
+from textual.widgets import Button, Input, Pretty, Static
+
+from chasten import constants
+
+CHECK_STORAGE = constants.chasten.App_Storage
+
+
+class ProperSentence:
+    def __init__(self, file_name: str):
+        """Initialize the ProperSentence object with a file name."""
+        self.file_name = file_name
+
+    def split_file(self) -> List[List[str]]:
+        """Split the file into a list of lists containing comma-separated values."""
+        check_list = []
+        with open(self.file_name) as file:
+            for row in file:
+                strip_row = row.strip()  # Remove leading/trailing white spaces
+                if strip_row:
+                    check_list.append(strip_row.split(","))
+        return check_list
+
+    def sentence_structure(self) -> str:
+        """Generate structured output based on the contents of the file."""
+        if os.path.isfile(CHECK_STORAGE):
+            check_list = self.split_file()
+            if len(check_list) != 0:
+                result = "Make a YAML file that checks for:"
+                for checks in check_list:
+                    quantity = "exactly" if checks[2] == "True" else "at least"
+                    result += f"\n - {quantity} {checks[1]} {checks[0]}"
+                return result
+            return "[red][ERROR][/red] No checks were supplied"
+        return f"[red][ERROR][/red] No {CHECK_STORAGE} file exists"
+
 
 # Define default values for checking and exact status
 Check = [
@@ -13,13 +48,11 @@ Check = [
     False,
 ]  # Check[0] stores "Check For:", Check[1] stores "Matches", Check[2] stores if you want exact checks
 Valid = False
-
 # Constants to map input field names to their positions in the Check list
 CHECK_VALUE = {
     "Check": 0,
     "Matches": 1,
 }
-C = "checks.txt"
 
 # Define input fields and buttons for the user interface
 Check_Input = Input(placeholder="Check For:", id="Check", name="Check")
@@ -32,21 +65,18 @@ Match_Input = Input(
 Exact_button = Button("Exact", id="Exact")  # Button to trigger an action
 
 
-class Bar(Static):
-    pass
-
-
 # Static widget to display user input and validation results
 class answers(Static):
-    def on_input_changed(self, event: Input.Changed):
+    def on_input_changed(self, event: Input.Changed) -> bool:
         """When inputs change this updates the values of Check"""
-        global Valid, Check
+        global Check, Valid  # noqa: PLW0602, PLW0603
         Valid = False
         if event.input.id == "Check":
             Check[CHECK_VALUE[event.input.name]] = event.input.value
         elif event.validation_result.is_valid:
             Check[CHECK_VALUE[event.input.name]] = event.input.value
             Valid = True
+        return Valid
 
     def compose(self) -> ComposeResult:
         """For displaying the user interface"""
@@ -58,14 +88,16 @@ class answers(Static):
 class button_prompts(Static):
     def on_button_pressed(self, event: Button.Pressed) -> None:
         """When a button is pressed this function determines what to do"""
-        global Valid, Check
+        global Check, Valid  # noqa: PLW0602, PLW0603
         if event.button.id == "Exact":
             Check[2] = True  # Mark the "Exact" button as clicked
             event.button.disabled = True  # Disable the "Exact" button after clicking
         elif event.button.id == "done":
-            app.exit()  # Exit the application if the "Done" button is clicked
+            config_App.exit(
+                self
+            )  # Exit the application if the "Done" button is clicked
         elif event.button.id == "clear":
-            with open(C, "w") as file:
+            with open(CHECK_STORAGE, "w") as file:
                 file.write(
                     ""
                 )  # Clears Checks.txt file when "Clear Check" button is clicked
@@ -78,17 +110,16 @@ class button_prompts(Static):
                 self.query_one(Pretty).update([])  # Clear any validation messages
                 Exact_button.disabled = False  # Re-enable the "Exact" button
                 Check_Input.value = ""
-                Match_Input.value = ""
-                app.refresh()  # Refresh the application UI
+                Match_Input.value = ""  # Refresh the application UI
         else:
             self.query_one(Pretty).update(["Invalid Input Please enter a Integer"])
             Match_Input.value = ""  # Clear the "Matches" input field
 
     def store_in_file(self):
         """Store inputed values into a text file"""
-        if not os.path.isfile("checks.txt"):
-            open(C, "x")  # Create a new file if it doesn't exist
-        with open(C, "a") as file:
+        if not os.path.isfile(CHECK_STORAGE):
+            open(CHECK_STORAGE, "x")  # Create a new file if it doesn't exist
+        with open(CHECK_STORAGE, "a") as file:
             file.write(
                 f"\n{Check[0]},{Check[1]},{Check[2]}"
             )  # Append input data to the file
@@ -97,7 +128,7 @@ class button_prompts(Static):
         """For displaying the user interface"""
         yield Pretty([])  # Widget to display validation messages
         yield Exact_button  # Display the "Exact" button
-        yield Button("Next Check!", id="next")  # Display the "Next Check!" button
+        yield Button("Submit Check!", id="next")  # Display the "Next Check!" button
         yield Button("Done!", id="done")  # Display the "Done!" button
         yield Button("Clear Checks", id="clear", variant="error")
 
@@ -141,12 +172,3 @@ class config_App(App):
         """For displaying the user interface"""
         yield answers()  # Display the input fields for user input
         yield button_prompts()  # Display the buttons for user interaction
-
-
-# Main entry point for the application
-if __name__ == "__main__":
-    app = config_App()
-    check = app.run()  # Run the Textual application
-    output = ProperSentence(C)
-    r = output.sentence_structure()
-    print(r)
